Finish migration of all code in the main SmSn read/write paths, using Hydra-generated data classes exclusively.

+++ smsn-models/src/gen-main/java/net/fortytwo/smsn/brain/Timestamp.java
@@ -5,16 +5,16 @@ package net.fortytwo.smsn.brain;
 import java.io.Serializable;
 
 /**
- * A Unix timestamp in seconds
+ * A Unix timestamp in milliseconds
  */
 public class Timestamp implements Serializable {
   public static final hydra.core.Name TYPE_NAME = new hydra.core.Name("net/fortytwo/smsn/brain.Timestamp");
   
   public static final hydra.core.Name FIELD_NAME_VALUE = new hydra.core.Name("value");
   
-  public final Integer value;
-  
-  public Timestamp (Integer value) {
+  public final Long value;
+
+  public Timestamp (Long value) {
     java.util.Objects.requireNonNull((value));
     this.value = value;
   }
